+++ tests/test_datasets/test_base_dataset.py
@@ -1,9 +1,9 @@
 import torch
 import pytest
 
-from torchchronos.datasets import BaseDataset
+from torchchronos.datasets.util.base_dataset import BaseDataset
 
-def test_Basedataset():
+def test_base_dataset():
     data = torch.randn(10, 1, 100)
     target = torch.randint(0, 2, (10,))
 
@@ -15,6 +15,5 @@ def test_Basedataset():
 
     dataset = BaseDataset(data)
     assert len(dataset) == 10
-    assert len(dataset[0]) == 2
-    assert dataset[0][0].shape == torch.Size([1, 100])
-    assert dataset[0][1] is None
+    assert len(dataset[0]) == 1
+    assert dataset[0].shape == torch.Size([1, 100])
